requirements.txt
wheel
nbresult
colorama
ipdb
ipykernel
yapf
matplotlib
pygeohash
pytest
seaborn
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.1
google-cloud-bigquery
google-cloud-storage
db-dtypes
pyarrow
# Trick to install the version of Tensorflow depending on your processor: darwin == Mac, ARM == M1
tensorflow-macos==2.10.0; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
tensorflow==2.10.0; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
tensorflow==2.10.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips
